FEAT: update and refactor data summary
